# generated/ecommerce-checkout/openapi.yaml
openapi: "3.0.3"
info:
  title: "ecommerce-checkout API"
  version: "1.0.0"
  description: "Generated API for ecommerce-checkout workflow"

servers:
  - url: http://localhost:8080
    description: Development server

paths:
  /health:
    get:
      summary: Health check
      operationId: getHealth
      tags:
        - system
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"

  /api/ecommercecheckout:
    post:
      summary: Create new ecommerce-checkout
      operationId: createEcommerceCheckout
      tags:
        - ecommercecheckout
      responses:
        "201":
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StateResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/ecommercecheckout/{id}:
    get:
      summary: Get ecommerce-checkout state
      operationId: getEcommerceCheckout
      tags:
        - ecommercecheckout
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Aggregate ID
      responses:
        "200":
          description: Current state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StateResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"


  /api/start_checkout:
    post:
      summary: "Begin checkout process"
      operationId: startCheckout
      tags:
        - transitions
      x-transition-id: start_checkout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionRequest"
      responses:
        "200":
          description: Transition fired successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransitionResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Transition failed (guard rejected or invalid state)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"


  /api/enter_payment:
    post:
      summary: "Enter payment details"
      operationId: enterPayment
      tags:
        - transitions
      x-transition-id: enter_payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionRequest"
      responses:
        "200":
          description: Transition fired successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransitionResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Transition failed (guard rejected or invalid state)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"


  /api/process_payment:
    post:
      summary: "Process the payment"
      operationId: processPayment
      tags:
        - transitions
      x-transition-id: process_payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionRequest"
      responses:
        "200":
          description: Transition fired successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransitionResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Transition failed (guard rejected or invalid state)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"


  /api/payment_success:
    post:
      summary: "Payment processed successfully"
      operationId: paymentSuccess
      tags:
        - transitions
      x-transition-id: payment_success
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionRequest"
      responses:
        "200":
          description: Transition fired successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransitionResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Transition failed (guard rejected or invalid state)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"


  /api/payment_fail_1:
    post:
      summary: "First payment attempt failed"
      operationId: paymentFail1
      tags:
        - transitions
      x-transition-id: payment_fail_1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionRequest"
      responses:
        "200":
          description: Transition fired successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransitionResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Transition failed (guard rejected or invalid state)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"


  /api/retry_payment_1:
    post:
      summary: "Retry payment (attempt 2)"
      operationId: retryPayment1
      tags:
        - transitions
      x-transition-id: retry_payment_1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionRequest"
      responses:
        "200":
          description: Transition fired successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransitionResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Transition failed (guard rejected or invalid state)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"


  /api/payment_fail_2:
    post:
      summary: "Second payment attempt failed"
      operationId: paymentFail2
      tags:
        - transitions
      x-transition-id: payment_fail_2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionRequest"
      responses:
        "200":
          description: Transition fired successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransitionResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Transition failed (guard rejected or invalid state)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"


  /api/retry_payment_2:
    post:
      summary: "Retry payment (attempt 3)"
      operationId: retryPayment2
      tags:
        - transitions
      x-transition-id: retry_payment_2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionRequest"
      responses:
        "200":
          description: Transition fired successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransitionResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Transition failed (guard rejected or invalid state)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"


  /api/payment_fail_3:
    post:
      summary: "Third payment attempt failed"
      operationId: paymentFail3
      tags:
        - transitions
      x-transition-id: payment_fail_3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionRequest"
      responses:
        "200":
          description: Transition fired successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransitionResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Transition failed (guard rejected or invalid state)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"


  /api/cancel_order:
    post:
      summary: "Cancel order after max retries"
      operationId: cancelOrder
      tags:
        - transitions
      x-transition-id: cancel_order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionRequest"
      responses:
        "200":
          description: Transition fired successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransitionResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Transition failed (guard rejected or invalid state)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"


  /api/fulfill:
    post:
      summary: "Fulfill the order"
      operationId: fulfill
      tags:
        - transitions
      x-transition-id: fulfill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionRequest"
      responses:
        "200":
          description: Transition fired successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransitionResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Transition failed (guard rejected or invalid state)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"


components:
  schemas:
    TransitionRequest:
      type: object
      required:
        - aggregate_id
      properties:
        aggregate_id:
          type: string
          description: ID of the aggregate to modify
        data:
          type: object
          description: Additional transition data
          additionalProperties: true

    TransitionResult:
      type: object
      properties:
        success:
          type: boolean
        aggregate_id:
          type: string
        version:
          type: integer
        state:
          type: object
          additionalProperties: true
        error:
          type: string

    StateResponse:
      type: object
      properties:
        aggregate_id:
          type: string
        version:
          type: integer
        state:
          $ref: "#/components/schemas/EcommerceCheckoutState"
        places:
          type: object
          additionalProperties:
            type: integer
          description: Current Petri net marking
        enabled_transitions:
          type: array
          items:
            type: string
          description: Transitions that can currently fire

    EcommerceCheckoutState:
      type: object
      description: Aggregate state for ecommerce-checkout
      properties:
        cart:
          type: integer
        checkout_started:
          type: integer
        payment_pending:
          type: integer
        payment_processing:
          type: integer
        retry_1:
          type: integer
        retry_2:
          type: integer
        retry_3:
          type: integer
        paid:
          type: integer
        cancelled:
          type: integer
        fulfilled:
          type: integer

    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

tags:
  - name: system
    description: System endpoints
  - name: ecommercecheckout
    description: ecommerce-checkout aggregate operations
  - name: transitions
    description: Petri net transition endpoints